+++ src/nationalarchives/stories/utilities/colour-schemes/colour-themes.stories.js
@@ -192,7 +192,7 @@ const Template = ({ theme, accent }) => {
         })}
       </div>
     </div>
-    <main id="main-content" role="main">
+    <main class="tna-main" id="main-content">
       ${Hero({
         params: {
           heading: "Title",
@@ -1244,6 +1244,11 @@ const Template = ({ theme, accent }) => {
             icon: "instagram",
             title: "The National Archives Instagram feed",
           },
+          {
+            href: "https://www.tiktok.com/@uknatarchives",
+            icon: "tiktok",
+            title: "The National Archives TikTok feed",
+          },
         ],
         navigation: [
           {
@@ -1314,6 +1319,10 @@ const Template = ({ theme, accent }) => {
               {
                 text: "The National Archives Trust",
                 href: "https://www.nationalarchives.gov.uk/about/the-national-archives-trust/",
+              },
+              {
+                text: "Friends of The National Archives",
+                href: "https://www.nationalarchives.gov.uk/about/get-involved/friends-of-the-national-archives/",
               },
             ],
           },