+++ tasks/global-header/generate-global-header-assets.js
@@ -94,6 +94,11 @@ const footerHTML = nunjucks
           icon: "instagram",
           title: "The National Archives Instagram feed",
         },
+        {
+          href: "https://www.tiktok.com/@uknatarchives",
+          icon: "tiktok",
+          title: "The National Archives TikTok feed",
+        },
       ],
       navigation: [
         {
@@ -164,6 +169,10 @@ const footerHTML = nunjucks
             {
               text: "The National Archives Trust",
               href: "https://www.nationalarchives.gov.uk/about/the-national-archives-trust/",
+            },
+            {
+              text: "Friends of The National Archives",
+              href: "https://www.nationalarchives.gov.uk/about/get-involved/friends-of-the-national-archives/",
             },
           ],
         },